feat: integrate JWT authentication and enhance API documentation
- Added JWT dependencies to pom.xml for token management.
- Implemented JwtInterceptor for request validation using JWT.
- Updated SwaggerConfig to support OpenAPI 3.0 and improved API documentation.
- Created WebConfig to define interceptor and whitelist paths.
- Enhanced MenuController and RoleMenuController to support tree structure and role-based menu retrieval.
- Updated UserController to include JWT token generation and user profile retrieval.
- Introduced new DTOs (LoginDTO, UserInfo) for better data handling.
- Refactored entity classes to include title and meta information for menus.
- Removed deprecated TokenUtils class and replaced it with JwtUtil for JWT operations.
- Updated application.yml for JWT configuration.

+++ SpringBoot/src/main/java/com/example/demo/entity/User.java
@@ -6,17 +6,13 @@ import com.baomidou.mybatisplus.annotation.TableId;
 import com.baomidou.mybatisplus.annotation.TableName;
 import lombok.Data;
 
-import org.apache.ibatis.annotations.Update;
-
 import javax.validation.constraints.NotEmpty;
-import javax.validation.constraints.NotNull;
-import java.util.Map;
 
 @TableName("user")
 @Data
 public class User {
     @TableId (type = IdType.AUTO)
-    private Integer id;
+    private Long id;
     @NotEmpty(message = "用户名不为空")
     private String username;
     private String nickName;
